stock_ETL - Copy/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
kafka-python
requests
pyspark
psycopg2-binary
apache-airflow-providers-http
